move Provider Baseclasses outside of the provider namespace to clean up
