+++ lib/Provider/Unsplash.php
@@ -21,7 +21,7 @@
  */
 
 namespace OCA\Unsplash\Provider;
-use OCA\Unsplash\Provider\Provider;
+use OCA\Unsplash\ProviderHandler\Provider;
 
 class Unsplash extends Provider{
