+++ lib/Provider/WallhavenCC.php
@@ -21,8 +21,8 @@
  */
 
 namespace OCA\Unsplash\Provider;
-use OCA\Unsplash\Provider\Provider;
 use OC\AppFramework\Http\Request;
+use OCA\Unsplash\ProviderHandler\Provider;
 
 class WallhavenCC extends Provider{
